pyproject.toml
[project]
name = "patchnote-prasia"
version = "0.1.0"
description = "프라시아 전기 패치노트 수집 및 검색 서비스"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "click>=8.1",
    "python-dotenv>=1.0",
    "numpy>=2.0",
    "scikit-learn>=1.8",
    "joblib>=1.5",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]
api = ["fastapi>=0.115", "uvicorn>=0.30"]

[project.scripts]
patchnote = "patchnote_prasia.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
